Add Node dotenv Package Dependency
The dotenv package allows easily creating a .env file in the project
root with needed environment variables. The .env file is ignored by Git
so that credentials and other secrets can be kept out of the repository.
The other benefit is that it will allow developers to create private
Cloudant databases to connect to while developing on the project.