Fix semantics of expected state in diff result
This is especially important in the creation of the Bolo where the
passed in bolo would not have an id. The behavior of the diff operation
only looks at enumerable keys on the source object. Having the original
bolo first means that the `id` key is not checked since it does not
exist in the set of enumerable keys.