Implement initial attachment retrieval in the web component
A route is added for attachments which proxies request to the service
layer.  The service layer proxies requests to the configures concrete
repository and returns data in a DTO much like the one used in the
creation of BOLOs with attachments.

Any bolo data passed to the UI layer can also check for attachments in
the `data.attachments` key.