Temporary fix for edit BOLO function.
The edit bolo form parsing is producing some kind of application/octet
binary file for some reason which is thus getting passed in as an
attachment. The binary being parsed does not have any field name
attached so the origin of the binary is unknown. As a temporary fix,
specific to this issue, only binary data with file names are added to
the attachment list.