Refactor create bolo
The creation of a bolo now takes three image uploads, a featured image
and two extra attachments. The bolo object stores these image
attachments meta data in an 'images' object that store uuid as the key
and the original filename as the value. This somewhat matches the way
cloudant stores attachments.